refactor(p1): extract ChatInputUsage.kt — context indicator + reasoning chip + usage sheet
First step on the ChatInput god class (3657 lines). Pulls the Usage/Reasoning
composer-row helpers out of ChatInput.kt into a sibling file, leaving the main
ChatInput composable untouched.

What moves to ChatInputUsage.kt (639 lines new):
- internal @Composable ContextUsageIndicator (called from ChatInput L625)
- internal @Composable ReasoningLevelChip (called from L653)
- internal @Composable ComposerUsageSheet (called from L521)
- internal fun OpenAICodexUsageStatus.toComposerUsageStatus(Context) (L366)
- internal fun ProviderUsageStatus.toComposerUsageStatus() (L368)
- internal fun ProviderSetting?.providerRoutingKey() (L518 + reasoningFamily)
- internal data class ComposerUsageStatus / ComposerUsageMetric (return types
  of the two internal mapper fns; data classes preserved for `==` parity)
- file-private: ReasoningLevelMenuCell, ComposerStatusChip, UsageMetricRow,
  Int.formatContextTokens, ProviderUsageMetric.toComposerUsageMetric,
  OpenAICodexUsageWindow.toComposerUsageMetric, Long.formatUsageResetDetail,
  ReasoningLevel.composerLabel, List<ReasoningOption>.labelFor,
  reasoningOptionsOf, ReasoningLevel.coerceToReasoningOptions,
  Model?.reasoningOptions, Model.reasoningFamily,
  String.isQwenPlusBinaryReasoningModel, ReasoningOption data class,
  ReasoningFamily enum

ChatInput.kt: 3657 → 3090 lines (-15%). Main ChatInput composable itself
untouched — this is a helper-only lift.

Visibility: 6 symbols `internal` (the call-graph boundary visible to the main
composable + provider-routing extension used in LaunchedEffect). Everything
else stays `private` inside the new file. ComposerUsageStatus / Metric promoted
to `internal data class` so they appear in the public signatures of the mappers.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL (2 pre-existing
Experimental Haze API warnings unchanged; same warnings present before extraction).
